python-node/requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
protobuf==5.27.2
pillow==11.0.0